Logica de levantar inmuebles con ID propietario realizada

+++ src/main/java/DAO/InmuebleDAO.java
@@ -13,9 +13,9 @@ public interface InmuebleDAO {
 
     public Inmueble getById(int id);
 
-    public List<Inmueble> listAllByPropietario(int idPropietario);
+    public List<Inmueble> listAllByPropietario(long idPropietario);
 
-    public List<Inmueble> listAllByPropietario(int idPropietario, int inicio, int fin);
+    public List<Inmueble> listAllByPropietario(long idPropietario, int inicio, int fin);
 
     public List<Inmueble> listAll();
